% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_lasso.R
\name{bayesian_lasso}
\alias{bayesian_lasso}
\title{Bayesian Lasso Regression}
\usage{
bayesian_lasso(
  df,
  outcome_var,
  protected_vars = NULL,
  save_output = FALSE,
  num.reps = 10000,
  a = 1,
  b = 0.1,
  burn = 501,
  seed = 123,
  output_name = "bayesian_lasso_output"
)
}
\arguments{
\item{df}{A data frame containing the variables for the analysis.}

\item{outcome_var}{The name of the outcome variable in the data frame.}

\item{protected_vars}{A vector of variable names that should not be penalized.}

\item{save_output}{Logical; if TRUE, the function saves the output in a specified location.}

\item{num.reps}{The number of MCMC repetitions.}

\item{a}{The shape parameter of the gamma distribution for the prior on lambda squared.}

\item{b}{The rate parameter of the gamma distribution for the prior on lambda squared.}

\item{burn}{The number of initial samples to discard (burn-in).}

\item{seed}{The random seed for reproducibility.}

\item{output_name}{The name of the output file to save the results.}
}
\value{
A list containing the lambda lasso, beta posterior, and mean beta.
}
\description{
Perform Bayesian Lasso regression by
protecting certain variables.
}
\examples{
n <- 100
x1 <- sample(c("A", "B"), n, replace = TRUE)
x1[sample(n, 10)] <- NA  # add missing values to the first column
x2 <- rnorm(n)
x2[sample(n, 5)] <- NA  # add missing values to the second column
x3 <- rnorm(n)
y <- as.numeric(x1 == "A") + x2 + rnorm(n)
mydata <- data.frame(x1, x2, x3, y)
model <- bayesian_lasso(df=mydata, outcome_var = "y", protected_vars = c("X1"))
}
