% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-boundaries.r
\name{calculate_boundaries}
\alias{calculate_boundaries}
\title{Calculate shared boundaries between planning units}
\usage{
calculate_boundaries(x, matrix, sparse, triangular, edge_factor)
}
\arguments{
\item{x}{Raster* or SpatialPolygons object; planning units for which
boundaries lengths are calculated.}

\item{matrix}{logical; whether to return a matrix (\code{TRUE}) or data frame
(\code{FALSE}).}

\item{sparse}{logical; whether to explicitly include shared boundaries of
length zero. If \code{sparse} and \code{matrix} and both \code{TRUE} then a
sparse matrix from the \code{slam} package is returned.}

\item{triangular}{logical; whether to return all matrix elements or just the
upper triangle. Since this matrix will typically be symmetric, the lower
triangle is redundant, and it may be desirable to drop it.}

\item{edge_factor}{numeric; factor to scale lengths of edge boundaries that
aren't shared with another planning unit. Set this to 0 if you don't wish
to include these boundaries as diagonal components of the matrix. This
scale factor can also be used in cases where edge boundaries are much
longer than other shared boundaries, such as when edges are composed of
complicated coast lines. In such situations, the long boundary lengths can
bias selection against these edge planning units, and it may be desirable
to scale the length to smaller values.}
}
\value{
If \code{matrix} is \code{TRUE} (the default), a matrix where each
  element corresponds to the length of the boundary between a pair of
  planning units. If \code{matrix} is \code{FALSE}, a data frame with three
  columns corresponding to the IDs of a pair of planning and the length of
  the shared boundary.

  Most pairs of planning units will have no shared boundary and by default
  these zero boundaries will not be stored explicitly. For matrices, this is
  accomplished using sparse matrix objects from the \code{slam} package,
  which use up much less memory than standard matrices. To explicitly include
  zero boundaries, use \code{sparse = FALSE}.
}
\description{
Prepare matrix of planning unit boundary lengths for a Marxan-like reseve
design exercise. Off diagonal components are shared boundaries between pairs
of planning units and diagonal components are external boundaries of planning
units, i.e. those not shared with another planning unit, such as boundaries
on the edge of the study area.
}
\details{
If planning units are provided as a SpatialPolygons object, it is recommended
that you install the \code{marxan} package (currently only available on
GitHub). This package has a boundary length function implemented in C++ that
will be called by \code{calculate_boundaries}. If \code{marxan} is not
installed, \code{calculate_boundaries} will use an alternate implementation,
which is several orders of magnitude slower.
}
\examples{
r <- raster::raster(matrix(1:9, 3, 3))
calculate_boundaries(r)
calculate_boundaries(r, sparse = FALSE, triangular = TRUE)
hex_grid <- make_grid(r, cell_width = 0.5, type = "hexagonal", clip = TRUE)
calculate_boundaries(hex_grid, matrix = FALSE)
}

