% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mse.R
\name{calculate_mse}
\alias{calculate_mse}
\title{Calculate MSE for Training and Testing Sets}
\usage{
calculate_mse(model, train, y, test)
}
\arguments{
\item{model}{A list containing the \code{mean_beta} vector which represents the model's coefficients.}

\item{train}{A dataframe representing the training set with all predictor variables and the outcome variable.}

\item{y}{The name of the outcome variable in the train and test datasets.}

\item{test}{A dataframe representing the testing set with all predictor variables and the outcome variable.}
}
\value{
A named vector with the MSE for the training set (\code{MSE_training}) and the testing set (\code{MSE_testing}).
}
\description{
This function computes the Mean Squared Error (MSE) for the training and testing datasets based on the specified model's coefficients.
}
\examples{
# NOT RUN
# calculate_mse(model = model1, train = train, test = test, y = "OutcomeVariableName")

}
