% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficients.R
\name{get_coefficients}
\alias{get_coefficients}
\title{Get Coefficients and Credible Intervals from BPL Model}
\usage{
get_coefficients(model, df, outcome_var, method = "HDI")
}
\arguments{
\item{model}{A list containing at least 'beta_post' and 'mean_beta', which are the posterior distributions and mean values of the coefficients, respectively.}

\item{df}{A dataframe containing the predictor variables and the outcome variable.}

\item{outcome_var}{The name of the outcome variable in the dataframe.}

\item{method}{The method used to calculate the credible intervals (defaults to "HDI" - Highest Density Interval).}
}
\value{
A dataframe with the names of the predictor variables, mean of the beta coefficients, lower and upper bounds of the credible intervals, and a logical flag indicating if zero is contained within the interval.
}
\description{
This function extracts the model coefficients and calculates their credible intervals, also checking if zero is contained within these intervals.
}
\examples{
n <- 50
x1 <- sample(c("A", "B"), n, replace = TRUE)
x1[sample(n, 10)] <- NA  # add missing values to the first column
x2 <- rnorm(n)
x2[sample(n, 5)] <- NA  # add missing values to the second column
x3 <- rnorm(n)
y <- as.numeric(x1 == "A") + x2 + rnorm(n)
mydata <- data.frame(x1, x2, x3, y)
model <- bayesian_lasso(df=mydata, outcome_var = "y", protected_vars = c("X1"))
get_coefficients(model, mydata, "y")

}
