% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_post.R
\name{plot_beta_post}
\alias{plot_beta_post}
\title{Plot Posterior Distributions for Model Coefficients}
\usage{
plot_beta_post(
  model,
  df,
  outcome_var,
  color = "blue",
  linewidth = 0.7,
  linetype = "dashed",
  method = "ETI"
)
}
\arguments{
\item{model}{A model object containing at least a \code{beta_post} matrix with posterior distributions of the coefficients.}

\item{df}{A dataframe containing the predictor variables and the outcome variable.}

\item{outcome_var}{The name of the outcome variable in the dataframe.}

\item{color}{The color to be used for points and lines in the plot.}

\item{linewidth}{The width of the line to be used in the plot.}

\item{linetype}{The type of the line to be used in the plot.}

\item{method}{The method used to calculate credible intervals (defaults to "ETI" - Equal-tailed Interval).}
}
\value{
A ggplot object representing the Bayesian Lasso Coefficients with their credible intervals.
}
\description{
This function takes a model object and a dataframe, identifies predictor variables, calculates credible intervals for coefficients, and produces a plot of these statistics.
}
\examples{
# Assuming 'model' is generated using bayesian_lasso function
# and 'df' is a dataframe with the required structure
# plot_beta_post(model = result1, df = mydata, outcome_var = "y")

}
