% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_from_coefs.R
\name{plot_from_coefs}
\alias{plot_from_coefs}
\title{Plot Coefficients from a Dataframe}
\usage{
plot_from_coefs(
  df,
  color = "blue",
  linewidth = 0.7,
  linetype = "dashed",
  ylim = NULL
)
}
\arguments{
\item{df}{A dataframe containing the names of the predictor variables and the corresponding statistics: mean_beta, lower_bound, upper_bound.}

\item{color}{The color to be used for points and lines in the plot.}

\item{linewidth}{The width of the line to be used in the plot.}

\item{linetype}{The type of the line to be used in the plot.}

\item{ylim}{Optional vector of length 2 to define the y-axis limits.}
}
\value{
A ggplot object representing the Bayesian Lasso Coefficients with their credible intervals.
}
\description{
This function takes a dataframe containing names, mean_beta, lower_bound, and upper_bound for model coefficients and produces a plot of these statistics.
}
\examples{
# Assuming 'df' is a dataframe produced by the `get_coefficients` function
# NOT RUN
# plot_from_coefs(df = coefficient_df)

}
