% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protectr-package.r
\docType{package}
\name{protectr}
\alias{protectr}
\alias{protectr-package}
\title{protectr: Systematic Protected Area Design in R}
\description{
Solve Marxan-like systematic reserve design problems using integer
programming techniques as implemented by the Gurobi Optimizer.
}

