% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Generate random numbers from a multivariate normal distribution. Borrowed from mvtnorm package on CRAN.}
\usage{
rmvnorm(
  n,
  mean = rep(0, nrow(sigma)),
  sigma = diag(length(mean)),
  method = c("eigen", "svd", "chol"),
  pre0.9_9994 = FALSE,
  checkSymmetry = TRUE
)
}
\arguments{
\item{n}{The number of samples to draw.}

\item{mean}{A numeric vector representing the means.}

\item{sigma}{A positive-definite symmetric matrix specifying the covariance matrix.}

\item{method}{A character string indicating which decomposition method to use.}

\item{pre0.9_9994}{Logical, for backwards compatibility.}

\item{checkSymmetry}{Logical, indicating if \code{sigma} should be checked for symmetry.}
}
\value{
A matrix of random numbers drawn from a multivariate normal distribution.
}
\description{
Generate random numbers from a multivariate normal distribution. Borrowed from mvtnorm package on CRAN.
}
\examples{
set.seed(123)
rmvnorm(5, mean = c(0, 0), sigma = matrix(c(1, 0, 0, 1), nrow = 2))
}
