% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-features.r
\name{summarize_features}
\alias{summarize_features}
\title{Calculate representation level of features in planning units}
\usage{
summarize_features(pu, features, matrix = TRUE, sparse = TRUE)
}
\arguments{
\item{pu}{SpatialPolygons* object; planning unit polygons.}

\item{features}{RasterStack object; distribution of features over study area
where each feature is a seperate layer in the RasterStack.}

\item{matrix}{logical; whether to return a matrix (\code{TRUE}) or data frame
(\code{FALSE}).s}

\item{sparse}{logical; whether to return a normal matrix or a sparse matrix
as implemented by the \code{slam} package. Sparse matrices only explicitly
store non-zero elements, which uses less memory.}
}
\value{
A matrix of feature representation levels within planning units where
  rows correspond to features and columns correspond to planning units. If
  \code{matrix = FALSE} a data frame is returned instead with columns
  corresponding to feature index, planning unit index, and amount within
  planning unit, repsectively.
}
\description{
Summarize a RasterStack of conservation features over a set of planning unit
polygons. If the \code{marxan} package is installed, this function is a
wrapper for \code{\link[marxan]{calcPuVsSpeciesData}}, otherwise a much
slower implementation is used that doesn't require \code{marxan}.
}
\examples{
# generate 4 feature distributions
e <- raster::extent(0, 100, 0, 100)
r <- raster::raster(e, nrows = 100, ncols = 100, vals = 1)
r_rare <- gaussian_field(r, range = 20, n = 2, prop = 0.1)
r_common <- gaussian_field(r, range = 20, n = 2, prop = 0.5)
r_features <- raster::stack(r_rare, r_common)
names(r_features) <- letters[1:4]

# create hexagonal grid
hex_grid <- make_grid(r, cell_area = 1000, type = "hexagonal")

# summarize
# spars matrix (default)
rij_sparse <- summarize_features(hex_grid, r_features)
rij_sparse
# normal matrix
rij_matrix <- summarize_features(hex_grid, r_features, sparse = FALSE)
dim(rij_matrix)
# data frame
rij_df <- summarize_features(hex_grid, r_features, matrix = FALSE)
head(rij_df)
}

